moved instance setup/selection to instancemanager.cpp comments
